Move EVMLogger to runtime module to break import cycle
