+++ state/tracer/tracer.go
@@ -6,7 +6,7 @@ import (
 	"math/big"
 	"time"
 
-	"github.com/dogechain-lab/dogechain/state/runtime/evm"
+	"github.com/dogechain-lab/dogechain/state/runtime"
 	"github.com/dogechain-lab/dogechain/types"
 )
 
@@ -21,7 +21,7 @@ type Context struct {
 // Tracer interface extends evm.EVMLogger and additionally
 // allows collecting the tracing result.
 type Tracer interface {
-	evm.EVMLogger
+	runtime.EVMLogger
 	GetResult() (json.RawMessage, error)
 	// Stop terminates execution of the tracer at the first opportune moment.
 	Stop(err error)
@@ -34,18 +34,18 @@ func NewDummyTracer() Tracer {
 	return &dummyTracer{}
 }
 
-func (d *dummyTracer) CaptureStart(txn evm.Txn, from, to types.Address, create bool,
+func (d *dummyTracer) CaptureStart(txn runtime.Txn, from, to types.Address, create bool,
 	input []byte, gas uint64, value *big.Int) {
 }
-func (d *dummyTracer) CaptureState(ctx evm.ScopeContext, pc uint64, op evm.OpCode,
+func (d *dummyTracer) CaptureState(ctx runtime.ScopeContext, pc uint64, op int,
 	gas, cost uint64, rData []byte, depth int, err error) {
 }
-func (d *dummyTracer) CaptureEnter(typ evm.OpCode, from, to types.Address,
+func (d *dummyTracer) CaptureEnter(typ int, from, to types.Address,
 	input []byte, gas uint64, value *big.Int) {
 }
 func (d *dummyTracer) CaptureExit(output []byte, gasUsed uint64, err error) {
 }
-func (d *dummyTracer) CaptureFault(ctx evm.ScopeContext, pc uint64, op evm.OpCode,
+func (d *dummyTracer) CaptureFault(ctx runtime.ScopeContext, pc uint64, op int,
 	gas, cost uint64, depth int, err error) {
 }
 func (d *dummyTracer) CaptureEnd(output []byte, gasUsed uint64, t time.Duration, err error) {
